services/api_frontend/requirements.txt
fastapi
uvicorn[standard]
aiokafka
motor
aioredis
pydantic
boto3
pyjwt
botocore
